Created Utility class to handle common functions
