+++ NavigatorApp/newCalculator/CollectionViewCell.swift
@@ -15,16 +15,8 @@ class CollectionViewCell: UICollectionViewCell {
         super.awakeFromNib()
     }
     
-    func generateRandomColor() -> UIColor {
-        let redValue = CGFloat(drand48())
-        let greenValue = CGFloat(drand48())
-        let blueValue = CGFloat(drand48())
-        let randomColor = UIColor(red: redValue, green: greenValue, blue: blueValue, alpha: 1.0)
-        return randomColor
-    }
-    
     func setUp(_ cellItems:[Int],_ index: Int){
-        self.backgroundColor = generateRandomColor()
+        self.backgroundColor = Utility.generateRandomColor()
         self.cellLabel.text = String(cellItems[index])
     }
